+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [1.16.0] - 2025-12-31
 ### Changed
 - Whether to print usage is optional when raising `ApplicationError` but
   default to no usage.
@@ -743,7 +745,8 @@ Major release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/util/compare/v1.15.14...HEAD
+[Unreleased]: https://github.com/plandes/util/compare/v1.16.0...HEAD
+[1.16.0]: https://github.com/plandes/util/compare/v1.15.14...v1.16.0
 [1.15.14]: https://github.com/plandes/util/compare/v1.15.13...v1.15.14
 [1.15.13]: https://github.com/plandes/util/compare/v1.15.12...v1.15.13
 [1.15.12]: https://github.com/plandes/util/compare/v1.15.11...v1.15.12